scripts/make_fit: Provide a way to add built modules
Provide arguments to support building a ramdisk from a directory tree of
modules. This is a convenient way to try out a kernel with its modules.

This makes use usr/gen_initramfs.sh where it can be found. Failing that
it uses the cpio tool rather than attempting to use a python module or
our own code. The list of modules is provided in a file.

Signed-off-by: Simon Glass <sjg@chromium.org>
Suggested-by: Ahmad Fatoum <a.fatoum@pengutronix.de>